refactor: migration of attachments and transcription functionality to the commands module
