+++ src/components/Assistant/FileList.tsx
@@ -6,8 +6,8 @@ import { useTranslation } from "react-i18next";
 
 import { useChatStore } from "@/stores/chatStore";
 import { useConnectStore } from "@/stores/connectStore";
-import { deleteAttachment, uploadAttachment } from "@/api/attachment";
 import FileIcon from "../Common/Icons/FileIcon";
+import { delete_attachment, upload_attachment } from "@/commands";
 
 interface FileListProps {
   sessionId: string;
@@ -39,7 +39,7 @@ const FileList = (props: FileListProps) => {
 
       if (uploaded) continue;
 
-      const attachmentIds = await uploadAttachment({
+      const attachmentIds = await upload_attachment({
         serverId,
         sessionId,
         filePaths: [path],
@@ -59,7 +59,7 @@ const FileList = (props: FileListProps) => {
   const deleteFile = async (id: string, attachmentId: string) => {
     setUploadFiles(uploadFiles.filter((file) => file.id !== id));
 
-    deleteAttachment({ serverId, id: attachmentId });
+    delete_attachment({ serverId, id: attachmentId });
   };
 
   return (
